fix: address PR review comments — correctness, config, and resilience
Resolves review threads:
  PRRT_kwDORTE6ps5vq2DV (products.rs: metadata overwrite)
  PRRT_kwDORTE6ps5vq2Di (main.rs: middleware ordering)
  PRRT_kwDORTE6ps5vq2DD (bills.rs: TOCTOU → ON CONFLICT DO NOTHING)
  PRRT_kwDORTE6ps5vq2DO (lib.rs: SELECT 1 i64→i32)
  PRRT_kwDORTE6ps5vq1WP (normalize.rs: history date ordering)
  PRRT_kwDORTE6ps5vq1WX (types.rs: SearchResult HashMap)
  PRRT_kwDORTE6ps5vq1Wr (sentiment/lib.rs: must_use stub)
  PRRT_kwDORTE6ps5vq1VX (brands.yaml: Green Monke shop_url null)
  PRRT_kwDORTE6ps5vq1VY (scbdb-cli/Cargo.toml: unused dep)
  PRRT_kwDORTE6ps5vq1V5 (config.rs: api_key_hash_salt optional)
  PRRT_kwDORTE6ps5vq1WG (seed.rs: is_active overwrite)
  PRRT_kwDORTE6ps5vq0kq (scbdb-core/Cargo.toml: dotenvy removed)
  PRRT_kwDORTE6ps5vq0lJ (scheduler.rs: stale comment)
  PRRT_kwDORTE6ps5vqylN (justfile: db-reset single-line)
  PRRT_kwDORTE6ps5vqylR (collect/mod.rs: pricing ? propagation)
  PRRT_kwDORTE6ps5vqylT (collect/mod.rs: products ? propagation)

Critical / correctness:
- lib.rs: fix SELECT 1 type annotation i64→i32 (Postgres int4 → i32)
- main.rs: fix Tower ServiceBuilder layer order; request_id must be
  outermost so spans capture the ID
- types.rs + client.rs: LegiScan search endpoint returns numbered objects
  {"0": ..., "1": ...}, not a Vec; use HashMap<String, Value> + flatten
- bills.rs + migration 400: replace non-atomic WHERE NOT EXISTS with
  unique index (NULLS NOT DISTINCT) + ON CONFLICT DO NOTHING

Resilience:
- collect/mod.rs: replace async-block + ? + match with loop-level match
  so an unexpected per-brand error logs and counts as a failure instead
  of aborting the entire collection run

Config:
- SCBDB_API_KEY_HASH_SALT: make optional (server-only concern, CLI
  commands don't require it); field type changed to Option<String>
- dotenvy: remove from scbdb-core (binary entrypoint concern only)

Minor:
- products.rs: remove metadata from DO UPDATE SET (hardcoded {} overwrites)
- seed.rs: remove is_active from DO UPDATE SET (re-activates manual deactivations)
- sentiment/lib.rs: remove #[must_use] from placeholder stub
- normalize.rs: use min()/max() over parsed dates (API order not guaranteed)
- brands.yaml: set Green Monke shop_url to null (Squarespace, not Shopify)
- scbdb-cli/Cargo.toml: remove unused scbdb-sentiment dep
- scheduler.rs: update stale phase comment
- justfile: db-reset single-line to prevent interactive prompt bypass

+++ crates/scbdb-db/src/lib.rs
@@ -115,7 +115,7 @@ pub async fn run_migrations(pool: &PgPool) -> Result<usize, sqlx::migrate::Migra
 ///
 /// Returns [`sqlx::Error`] if the query fails.
 pub async fn ping(pool: &PgPool) -> Result<(), sqlx::Error> {
-    sqlx::query_scalar::<_, i64>("SELECT 1")
+    sqlx::query_scalar::<_, i32>("SELECT 1")
         .fetch_one(pool)
         .await?;
     Ok(())
